requirements.txt
helper==2.5.0
invoke==2.2.0
matplotlib==3.10.0
scikit_learn==1.6.0
torch==2.5.1
tqdm==4.67.1
typer==0.15.1
hydra-core==1.3.2
